+++ game/src/game/game_state_manager.ts
@@ -1,3 +1,4 @@
+import { ReactStateSetters } from './react_state_setters';
 import { ServerConnection } from './server/connection/server_connection';
 import { ServerList } from './server/server_list';
 import { WarServer } from './server/war_server';
@@ -5,7 +6,7 @@ import { WarServer } from './server/war_server';
 export class GameStateManager {
     private _serverList = new ServerList();
     private _currentServer?: WarServer;
-    private _serverConnectionChangeListeners: ((connection?: WarServer) => void)[] = [];
+    readonly reactState = new ReactStateSetters();
 
     async initialize() {
         await this._serverList.initializeDB(game.engine.db);
@@ -17,11 +18,7 @@ export class GameStateManager {
 
     connectToServer(target: ServerConnection) {
         this._currentServer = new WarServer(target);
-        this._serverConnectionChangeListeners.forEach(l => l(this._currentServer));
-    }
-
-    onServerConnectionChange(listener: (connection?: WarServer) => void) {
-        this._serverConnectionChangeListeners.push(listener);
+        this.reactState.useGameSession.setConnection(target);
     }
 
     get serverList() {