+++ game/src/game/lobby/lobby_chat.ts
@@ -1,11 +1,9 @@
-import { Listener } from 'typeUtils';
 import { ClientPacketSendChatMessage } from '../server/connection/packet/to_send/lobby/common/send_chat_message';
 
 export type LobbyChatMessage = { sender: string; msg: string };
 
 export class LobbyChat {
     private _messages: LobbyChatMessage[] = [];
-    private _listeners: Listener<LobbyChatMessage[]>[] = [];
 
     sendMessage(msg: string) {
         new ClientPacketSendChatMessage(msg).dispatch();
@@ -13,20 +11,12 @@ export class LobbyChat {
 
     registerNewMessage(msg: LobbyChatMessage) {
         this._messages.push(msg);
-        this.triggerListeners();
+        game.state.reactState.useGameSession.setChat([...this._messages]);
     }
 
     eraseHistory() {
         this._messages = [];
-        this.triggerListeners();
-    }
-
-    private triggerListeners() {
-        this._listeners.forEach(l => l(this._messages));
-    }
-
-    onUpdate(listener: Listener<LobbyChatMessage[]>) {
-        this._listeners.push(listener);
+        game.state.reactState.useGameSession.setChat([]);
     }
 
     get messages() {