+++ game/src/game/server/connection/packet/listeners/game/pl_game_connection_token.ts
@@ -3,9 +3,8 @@ import { PacketListener } from '../packet_listener';
 export class PLGameConnectionToken extends PacketListener {
     register(): void {
         this.socket.on('gGameSessionConnectionToken', token => {
-            if (!this.server.currentLobby.value) return;
-            if (game.state.server?.currentLobby.value?.gameSession.value)
-                game.state.server.currentLobby.value.gameSession.value.token.value = token;
+            if (!this.server.currentLobby) return;
+            if (game.state.server?.currentLobby?.gameSession) game.state.server.currentLobby.gameSession.token = token;
         });
     }
 }