+++ game/src/game/server/connection/packet/listeners/game/pl_initial_game_state.ts
@@ -4,8 +4,8 @@ import { PacketListener } from '../packet_listener';
 export class PLInitialGameState extends PacketListener {
     register(): void {
         this.socket.on('gInitialGameState', state => {
-            if (!this.server.currentLobby.value) return;
-            this.server.currentLobby.value.gameSession.value = new WarGameSession(state);
+            if (!this.server.currentLobby) return;
+            this.server.currentLobby.gameSession = new WarGameSession(state);
         });
     }
 }