+++ game/src/game/server/connection/packet/listeners/lobby/pl_chat.ts
@@ -3,8 +3,8 @@ import { PacketListener } from '../packet_listener';
 export class PLChat extends PacketListener {
     register(): void {
         this.socket.on('chatMessage', (sender, msg) => {
-            if (this.server.currentLobby.value) {
-                this.server.currentLobby.value.chat.registerNewMessage({ sender, msg });
+            if (this.server.currentLobby) {
+                this.server.currentLobby.chat.registerNewMessage({ sender, msg });
             }
         });
     }