+++ game/src/game/server/connection/packet/listeners/lobby/pl_game_start_cancelled.ts
@@ -3,7 +3,7 @@ import { PacketListener } from '../packet_listener';
 export class PLGameStartCancelled extends PacketListener {
     register(): void {
         this.socket.on('lGameStartCancelled', () => {
-            game.state.server?.currentLobby.value?.cancelGameStartCountdown();
+            game.state.server?.currentLobby?.cancelGameStartCountdown();
         });
     }
 }