+++ game/src/game/server/connection/packet/listeners/lobby/pl_joined_lobby.ts
@@ -4,7 +4,7 @@ import { PacketListener } from '../packet_listener';
 export class PLJoinedLobby extends PacketListener {
     register(): void {
         this.socket.on('joinedLobby', pkt => {
-            this.server.currentLobby.value = new WarGameLobby(pkt);
+            this.server.currentLobby = new WarGameLobby(pkt);
         });
     }
 }