+++ game/src/game/server/connection/packet/listeners/lobby/pl_left_lobby.ts
@@ -3,8 +3,8 @@ import { PacketListener } from '../packet_listener';
 export class PLLeftLobby extends PacketListener {
     register(): void {
         this.socket.on('leftLobby', kicked => {
-            this.server.currentLobby.value = undefined;
-            this.server.lastLobbyExitReason.value = kicked ? 'kicked' : 'left';
+            this.server.currentLobby = undefined;
+            this.server.lastLobbyExitReason = kicked ? 'kicked' : 'left';
         });
     }
 }