+++ game/src/game/server/connection/packet/listeners/lobby/pl_starting_game.ts
@@ -3,7 +3,7 @@ import { PacketListener } from '../packet_listener';
 export class PLStartingGame extends PacketListener {
     register(): void {
         this.socket.on('lStartingGame', countdown => {
-            game.state.server?.currentLobby.value?.setGameStartingCountdown(countdown);
+            game.state.server?.currentLobby?.setGameStartingCountdown(countdown);
         });
     }
 }