+++ game/src/game/server/connection/packet/listeners/lobby/pl_update_lobby_state.ts
@@ -3,8 +3,8 @@ import { PacketListener } from '../packet_listener';
 export class PLUpdateLobbyState extends PacketListener {
     register(): void {
         this.socket.on('updateLobbyState', pkt => {
-            if (this.server.currentLobby.value) {
-                this.server.currentLobby.value.state.value = pkt;
+            if (this.server.currentLobby) {
+                this.server.currentLobby.state = pkt;
             } else {
                 console.warn('Trying to set lobby state for non-existent lobby');
             }