+++ game/src/hooks/use_game.tsx
@@ -1,4 +1,4 @@
-import React, { createContext, useContext, useMemo, useState } from 'react';
+import React, { createContext, useContext, useEffect, useMemo, useState } from 'react';
 import { WarGame } from '../game/war_game';
 
 interface IGameContext {
@@ -10,6 +10,11 @@ const GameContext = createContext<IGameContext>({} as IGameContext);
 
 const GameProvider: React.FC<{ children?: React.ReactNode }> = ({ children }) => {
     const [gameInstance, setGameInstance] = useState<WarGame>();
+
+    useEffect(() => {
+        if (!gameInstance) return;
+        gameInstance.state.reactState.useGame.setGameInstance = setGameInstance;
+    }, [gameInstance, setGameInstance]);
 
     const gameMemo = useMemo<IGameContext>(() => {
         return {