+++ CHANGELOG.md
@@ -15,6 +15,7 @@ A lot of progress tracking was lost because I didn't think to do this sooner.
 - 🖥 Reworked the server-side game handling logic to be better structured.
 - 🎮 Updated all game hooks to use `memo`s.
 - 🎮 The game will now always try to ping a server when connecting to it, even if the ping was already successful.
+- 🎮 Reworked client-side logic to directly call react state setters.
 
 ## [0.0.3]